phase-III/todo-ai-chatbot/todo-ai-chatbot/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn==0.24.0
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
sqlalchemy==2.0.23
uuid==1.30